tests: disable code sensitive to math library accuracy
The cos function here is supposed to output 0, but due to the values
used, never quite gets there.  Depending on the accuracy of the math
library, just how close it gets can change, so disable the test for
now.  Can revisit if pcalc ever switches to something like GMP.

Fixes #6.